显示点信息数目成功

+++ Map/Classes/MAPHomePage/Model/MAPGetPointModel.h
@@ -18,10 +18,10 @@
 @property (nonatomic, copy) NSString *createAt;
 @property (nonatomic, assign) int createBy;
 @property (nonatomic, assign) int ID;
-@property (nonatomic, assign) int mesCount;
-@property (nonatomic, assign) int phoCount;
-@property (nonatomic, assign) int audCount;
-@property (nonatomic, assign) int vidCount;
+@property (nonatomic, assign) long mesCount;
+@property (nonatomic, assign) long phoCount;
+@property (nonatomic, assign) long audCount;
+@property (nonatomic, assign) long vidCount;
 @property (nonatomic, assign) double latitude;
 @property (nonatomic, assign) double longitude;
 @property (nonatomic, copy) NSString *pointName;
